+++ plugin-agave/src/config.rs
@@ -63,8 +63,6 @@ pub struct ConfigChannel {
     #[serde(deserialize_with = "deserialize_num_str")]
     pub max_messages: usize,
     #[serde(deserialize_with = "deserialize_num_str")]
-    pub max_slots: usize,
-    #[serde(deserialize_with = "deserialize_num_str")]
     pub max_bytes: usize,
 }
 
@@ -72,9 +70,8 @@ impl Default for ConfigChannel {
     fn default() -> Self {
         Self {
             encoder: ProtobufEncoder::Raw,
-            max_messages: 2_097_152, // assume 20k messages per slot, aligned to power of 2
-            max_slots: 100,
-            max_bytes: 10 * 1024 * 1024 * 1024, // 10GiB, assume 100MiB per slot
+            max_messages: 2_097_152, // aligned to power of 2, ~20k/slot should give us ~100 slots
+            max_bytes: 15 * 1024 * 1024 * 1024, // 15GiB with ~150MiB/slot should give us ~100 slots
         }
     }
 }