+++ CHANGELOG.md
@@ -26,6 +26,8 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
+- plugin-agave,richat: remove `max_slots` ([#68](https://github.com/lamports-dev/richat/pull/68))
+
 ## 2025-01-22
 
 - filter-v2.2.0